neofitness-api/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
argon2-cffi==23.1.0
PyJWT==2.9.0
email-validator==2.2.0
